httpsrv: Clean up client state transitions; fix problems with pipelining.
The problem with pipelining was due to parsing of the subsequent request
if it was already (partially) read into the read_buf, while something
was still happening with the first request.  The http-parser needs to be
paused.

This was not caught earlier because we were previously only reading one
byte into the read_buf at a time, for testing.